requirements.txt
annotated-types==0.7.0
biopython==1.85
blis==0.7.11
catalogue==2.0.10
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpathlib==0.23.0
confection==0.1.5
conllu==6.0.0
cymem==2.0.11
en-ner-bc5cdr-md @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz
idna==3.11
jinja2==3.1.6
joblib==1.5.2
langcodes==3.5.0
language-data==1.3.0
loguru==0.7.3
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
nmslib-metabrainz==2.1.3
numpy==1.26.4
packaging==25.0
preshed==3.0.10
psutil==7.1.1
pybind11==3.0.1
pydantic==2.12.3
pydantic-core==2.41.4
pygments==2.19.2
pysbd==0.3.4
requests==2.32.5
rich==14.2.0
ruff==0.14.1
scikit-learn==1.7.2
scipy==1.16.2
scispacy==0.6.2
setuptools==80.9.0
shellingham==1.5.4
smart-open==7.4.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.2.5
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==2.0.0
